+++ dwds/lib/src/handlers/dev_handler.dart
@@ -179,7 +179,7 @@ class DevHandler {
 
     var webkitDebugger = WebkitDebugger(WipDebugger(tabConnection));
 
-    var metadataProvider = FileMetadataProvider(_assetReader);
+    var metadataProvider = FileMetadataProvider(_assetReader, _logWriter);
 
     return DebugService.start(
       // We assume the user will connect to the debug service on the same
@@ -458,7 +458,7 @@ class DevHandler {
       }
       var appId = devToolsRequest.appId;
       if (_servicesByAppId[appId] == null) {
-        var metadataProvider = FileMetadataProvider(_assetReader);
+        var metadataProvider = FileMetadataProvider(_assetReader, _logWriter);
 
         var debugService = await DebugService.start(
             _hostname,